% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATS_peak_detection.R
\name{ATS_peak_detection}
\alias{ATS_peak_detection}
\title{Apply Basic Cues Function to a Specific Site}
\usage{
ATS_peak_detection(
  Results_days,
  siteforsub,
  bio_data_all,
  climate.path,
  refday,
  lastdays,
  rollwin,
  lag,
  threshold
)
}
\arguments{
\item{Results_days}{A data frame containing the results method for the site, including slope and R-squared values.}

\item{siteforsub}{String. The unique identifier for the site (e.g., longitude and latitude).}

\item{climate.path}{String. The path to the directory containing the climate data for the site.}

\item{refday}{Integer. The reference day used for the analysis (default is 305).}

\item{lastdays}{Integer. The total number of days used for the analysis (default is 600).}

\item{rollwin}{Integer. The size of the rolling window for climate data smoothing (default is 1).}

\item{bio.data}{A data frame containing seed production data, with site-specific information such as \code{plotname.lon.lat} and \code{Year}.}
}
\value{
A data frame summarizing the fitted model for each identified weather cue window. If no signals are detected, the function will return a message indicating that no windows were identified.
}
\description{
This function applies the \code{runing_basic_cues} method to analyze climate and seed data for a specific site. It identifies weather cues based on the provided results from a CSP analysis and climate data for the site.
}
\details{
The function filters the seed data for the specified site and applies the \code{runing_basic_cues} function to analyze the relationship between climate variables and biological data. It identifies key weather cues by applying a thresholding algorithm to the results of a CSP analysis.
}
\examples{
\dontrun{
# Example usage:
cues_summary <- ATS_peak_detection(Results_days = Results_days,
                                        siteforsub = "longitude=-0.15_latitude=50.85",
                                        Fagus.seed = Fagus.seed,
                                        climate.beech.path = "path/to/climate/data",
                                        refday = 305,
                                        lastdays = 600,
                                        rollwin = 1)
}

}
\seealso{
\code{\link{runing_basic_cues}}
}
