% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correlation.spearman.se}
\alias{correlation.spearman.se}
\title{Calculate Standard Error of Spearman Correlation}
\usage{
correlation.spearman.se(cor, n)
}
\arguments{
\item{cor}{A numeric value representing the Spearman correlation coefficient.}

\item{n}{An integer representing the sample size.}
}
\value{
A numeric value representing the standard error of the Spearman correlation coefficient.
}
\description{
This function computes the standard error (SE) of Spearman's rank correlation coefficient
based on the sample size and the correlation coefficient.
}
\details{
The standard error is calculated using the formula:
\deqn{ \text{SE}_{\text{cor}} = \sqrt{\frac{(1 - \text{cor}^2)^2 (1 + \text{cor}^2 / 2)}{n - 3}} }
where \code{cor} is the Spearman correlation coefficient and \code{n} is the sample size.
This formula is used to estimate the variability of the Spearman correlation coefficient.
}
\examples{
# Example usage of the correlation.spearman.se function
spearman_cor <- 0.8
sample_size <- 50
se <- correlation.spearman.se(spearman_cor, sample_size)
print(se)

}
