% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_blocks}
\alias{create_blocks}
\title{Create Non-Overlapping Data Blocks}
\usage{
create_blocks(data, num_blocks)
}
\arguments{
\item{data}{A data frame or tibble to be divided into blocks. The data must be
arranged such that the order of the rows reflects the desired
sequential structure (e.g., time series data).}

\item{num_blocks}{An integer specifying the number of non-overlapping blocks to create
from the dataset. Must be a positive integer.}
}
\value{
A list containing the specified number of blocks, where each block is a
data frame. If the specified number of blocks cannot be created from the
dataset, an error is raised.
}
\description{
This function divides a given dataset into a specified number of non-overlapping blocks.
Each block contains a contiguous subset of rows from the dataset. The number of rows in
each block is determined by the total number of rows divided by the number of blocks.
}
\examples{
# Example dataset
example_data <- data.frame(year = 2000:2029, value = rnorm(30))

# Create 5 non-overlapping blocks from the dataset
blocks <- create_blocks(example_data, num_blocks = 5)

}
