% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_consecutive_sequences}
\alias{extract_consecutive_sequences}
\title{Extract Consecutive Sequences from a Vector}
\usage{
extract_consecutive_sequences(values, keep_all = FALSE)
}
\arguments{
\item{values}{Numeric vector. A vector of numeric values from which to extract consecutive sequences.}

\item{keep_all}{Logical. If \code{TRUE}, the function returns all sequences of consecutive numbers found in the input vector. If \code{FALSE}, it returns only the longest sequence. Default is \code{FALSE}.}
}
\value{
If \code{keep_all} is \code{FALSE}, returns a numeric vector of the longest sequence of consecutive numbers. If \code{keep_all} is \code{TRUE}, returns a list where each element is a numeric vector representing a sequence of consecutive numbers.
}
\description{
This function identifies and extracts sequences of consecutive numbers from a given vector. It can either return all sequences of consecutive numbers or only the longest sequence.
}
\details{
The function iterates through the numeric vector to find sequences where each value is consecutive (i.e., each value is exactly one greater than the previous value). It creates a list of these sequences. Depending on the \code{keep_all} parameter, it either returns all found sequences or just the longest one.
}
\examples{
# Example usage of extract_consecutive_sequences
values <- c(1, 2, 3, 5, 6, 7, 10, 11, 12)
longest_sequence <- extract_consecutive_sequences(values, keep_all = FALSE)
all_sequences <- extract_consecutive_sequences(values, keep_all = TRUE)
print(longest_sequence) # Prints: [1, 2, 3]
print(all_sequences)    # Prints: [[1, 2, 3], [5, 6, 7], [10, 11, 12]]

}
