% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL.moving.climate.analysis.R
\name{FULL.moving.climate.analysis}
\alias{FULL.moving.climate.analysis}
\title{Main Function of correlation - regression and process accross all sites}
\usage{
FULL.moving.climate.analysis(
  seed.data.all,
  climate.path,
  refday = 305,
  lastdays = 700,
  rollwin = 1,
  myform = formula("log.seed ~ TMEAN"),
  model_type = "lm"
)
}
\arguments{
\item{climate.path}{A string specifying the path to the directory containing climate data files. These files should be named according to the site names.}

\item{refday}{An integer representing the reference day for the rolling window analysis. Default is 305.}

\item{lastdays}{An integer specifying the number of days to include in the rolling window analysis. Default is the maximum of a specified range.}

\item{rollwin}{An integer specifying the rolling window size. Default is 1.}

\item{seed.data}{A data frame containing seed count and site-level information. It should include a column \code{plotname.lon.lat} to specify site names and a column \code{log.seed} for seed counts.}
}
\value{
A data frame containing the results of the moving window analysis for all sites. Includes correlation coefficients, model coefficients, and other relevant statistics for each site.
}
\description{
This function applies the moving climate analysis to all sites listed in the \code{seed.data} data frame. It iterates over each site, performs the moving window analysis, and combines the results into a single data frame.
}
\details{
The function performs the following steps:
\itemize{
\item Identifies unique site names from the \code{seed.data} data frame.
\item Applies the \code{site.moving.climate.analysis} function to each site using the unique site names.
\item Combines the results from all sites into a single data frame.
}
}
\examples{
# Example usage:
Fagus_seed <- data.frame(
  plotname.lon.lat = rep(c('Site1', 'Site2'), each = 10),
  log.seed = rnorm(20),
  year = rep(2000:2009, 2)
)
climate_beech_path <- 'path/to/climate/data'
result <- FULL.moving.climate.analysis(
  seed.data = Fagus_seed,
  climate.path = climate_beech_path,
  refday = 305,
  lastdays = 30,
  rollwin = 1
)

}
