% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.mean.sd}
\alias{get.mean.sd}
\title{Get Mean and Standard Deviation of a Vector}
\usage{
get.mean.sd(vector)
}
\arguments{
\item{vector}{Numeric vector. The input data from which the mean and standard deviation will be computed.}
}
\value{
A numeric vector of length two containing:
\itemize{
\item The mean value (first element).
\item The standard deviation (second element).
}
}
\description{
This function computes the mean and standard deviation of a numeric vector.
}
\details{
This function returns the average (mean) and standard deviation of the input numeric vector.
If the vector contains missing values (\code{NA}), they will be ignored in the calculation.
}
\examples{
vec <- c(1, 2, 3, 4, 5)
get.mean.sd(vec) # returns c(3, 1.581139)

vec_with_na <- c(NA, 2, 3, 4, 5)
get.mean.sd(vec_with_na) # returns c(3.5, 1.290994)

}
