% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.min.max}
\alias{get.min.max}
\title{Get Minimum and Maximum of a Vector}
\usage{
get.min.max(vector)
}
\arguments{
\item{vector}{Numeric vector. The input data from which the minimum and maximum values will be computed.}
}
\value{
A numeric vector of length two containing:
\itemize{
\item The minimum value (first element).
\item The maximum value (second element).
}
}
\description{
This function computes the minimum and maximum values of a numeric vector.
}
\details{
This function returns the smallest and largest values from the input numeric vector.
If the vector contains missing values (\code{NA}), they will be ignored in the calculation.
}
\examples{
vec <- c(1, 2, 3, 4, 5)
get.min.max(vec) # returns c(1, 5)

vec_with_na <- c(NA, 2, 3, 4, 5)
get.min.max(vec_with_na) # returns c(2, 5)

}
