% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{goingbackpastdayscalendar}
\alias{goingbackpastdayscalendar}
\title{Generate Calendar for "Historical" Days Reference}
\usage{
goingbackpastdayscalendar(refday = 274, lastdays = 1095, yearback = 3)
}
\arguments{
\item{refday}{Numeric. The reference day of the year (DOY) from which to start the backward calculation. Default is 274.}

\item{lastdays}{Numeric. The total number of days to go back into the past from the reference day. Default is 1095 (1095days, >3 years).}

\item{yearback}{Numeric. The number of years to go back from the reference year. Default is 3 years.}
}
\value{
A data frame containing:
\item{DATE}{The date of each day in the sequence.}
\item{YEAR}{The year of each date.}
\item{MONTHab}{The abbreviated month name.}
\item{DOY}{The Day of Year (DOY) for each date.}
\item{days.reversed}{The number of days going back from the reference day.}
}
\description{
This function generates a calendar to map days to their corresponding Day of Year (DOY)
for a specified number of days going back into the past from a reference day. It helps in
creating a historical reference for time series analysis by aligning days from past years.
For example if a day reversed starting from 1st november is 10, it means it is 10 days before
1st of November but will provide here the DOY
The function is not best, but still working
meaning that you might need to adjust this function, because here I want this mostly for my figure
}
\details{
The function creates a calendar by generating dates from a base period (1946 to 1953) and maps these dates to their corresponding DOY. It then calculates the days going back from a specified reference day and creates a sequence of dates accordingly. The function ensures that the number of days to go back and the number of years specified are consistent.

The function will stop and return an error if \code{lastdays} is greater than 365 and \code{yearback} is less than 2, or if \code{lastdays} is greater than 365*2 and \code{yearback} is less than 3.
}
\examples{
# Example usage of the goingbackpastdayscalendar function
calendar_data <- goingbackpastdayscalendar(refday = 274, lastdays = 1095, yearback = 3)
head(calendar_data)

}
