% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reformat.climate.backtothepast}
\alias{reformat.climate.backtothepast}
\title{Format Climate Data for Moving Windows analysis}
\usage{
reformat.climate.backtothepast(
  yearsref = 2000,
  climate = climate,
  yearneed = 2,
  refday = 274,
  lastdays = 1095,
  rollwin = 1,
  variablemoving = "temperature.degree"
)
}
\arguments{
\item{yearsref}{Numeric. The reference year from which to start the calculations. Default is 2000.}

\item{climate}{Data frame. The climate dataset containing variables such as temperature and date information.}

\item{yearneed}{Numeric. The number of years needed to adjust for fruit maturation time. Default is 2. But the idea is that if you want to look more than 1095 days to the past, you need to adjust here}

\item{refday}{Numeric. The reference day of the year (DOY) from which the moving window calculation begins. Default is 274.}

\item{lastdays}{Numeric. The total number of days to go back from the reference day. Default is 1095 (3 years).}

\item{rollwin}{Numeric. The window size for the rolling average calculation. Default is 1.}

\item{variablemoving}{Character. The name of the climate variable to apply the rolling average to. Default is 'temperature.degree' because it is my variable temperature name.}
}
\value{
A data frame containing:
\item{LONGITUDE}{Longitude of the location.}
\item{LATITUDE}{Latitude of the location.}
\item{year}{The year of each record.}
\item{date}{The date of each record.}
\item{yday}{Day of Year (DOY) for each date.}
\item{days.reversed}{Number of days going back from the reference day.}
\item{rolling_avg_tmean}{Rolling average of the specified climate variable.}
}
\description{
This function processes climate data to create a historical moving window of climate variables based on a reference day and a period going back in time.
It adjusts for leap years, applies a rolling average to the specified climate variable, and prepares the data for analysis over past periods.
Here you could adjust the rolling window to more than 1day (for example if you want average temperature over 7 days, or more or less), but in our method comparison, we set the default to 1
}
\details{
The function performs the following tasks:
\itemize{
\item Checks if the specified variable exists in the climate dataset.
\item Adjusts the reference day if the reference year is a leap year.
\item Filters the climate data to include only the relevant years.
\item Creates a sequence of days going back from the reference day.
\item Applies a rolling average to the specified climate variable.
\item Returns a data frame with the processed climate data, including the rolling average.
}

If the specified \code{variablemoving} is not found in the climate dataset, a warning is issued, and the function returns \code{NULL}.
}
\examples{
# Example usage of the reformat.climate.backtothepast function
processed_climate_data <- reformat.climate.backtothepast(yearsref = 2000, 
                                                         climate = climate_data, 
                                                         yearneed = 2, 
                                                         refday = 274, 
                                                         lastdays = 1095, 
                                                         rollwin = 7, 
                                                         variablemoving = 'temperature.degree')
head(processed_climate_data)
 
}
