% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns_if_needed.R
\name{rename_columns_if_needed}
\alias{rename_columns_if_needed}
\title{Rename Climwin Output Columns If Present}
\usage{
rename_columns_if_needed(df)
}
\arguments{
\item{df}{A data frame potentially containing the columns \code{WindowOpen} and \code{WindowClose}.}
}
\value{
A data frame with renamed columns if applicable.
}
\description{
This helper function checks whether the columns \code{WindowOpen} and \code{WindowClose} exist in a data frame
and renames them to \code{window.open} and \code{window.close}, respectively. It is useful for standardizing
output from the \code{climwin} package.
}
\examples{
df <- data.frame(WindowOpen = 100, WindowClose = 200)
rename_columns_if_needed(df)

}
