% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_signal.R
\name{replace_0}
\alias{replace_0}
\title{Replace Sequences of Zeros in a Vector with Neighboring Values}
\usage{
replace_0(vec, threshold)
}
\arguments{
\item{vec}{A numeric vector that may contain sequences of zeros.}

\item{threshold}{An integer specifying the maximum length of consecutive zeros to replace. If the length of a zero sequence is less than or equal to \code{threshold}, the zeros are replaced by the nearest non-zero value. Sequences longer than the threshold are not modified.}
}
\value{
A modified version of the input vector, where certain sequences of zeros have been replaced by nearby non-zero values.
}
\description{
This function replaces sequences of zeros in a vector with the value of the closest non-zero element, provided the sequence length is within a defined threshold.
}
\details{
The function scans the input vector for sequences of consecutive zeros and replaces them with the nearest preceding or following non-zero value, if the length of the zero sequence is within the specified \code{threshold}. Sequences longer than \code{threshold} are left unchanged.
}
\examples{
vec <- c(1, 2, 0, 0, 3, 4, 0, 0, 0, 5)
replace_0(vec, threshold = 2)
# Expected output: c(1, 2, 2, 2, 3, 4, 0, 0, 0, 5)

}
