% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runing.movingwin.analysis}
\alias{runing.movingwin.analysis}
\title{Perform Moving Window Analysis on Climate Data}
\usage{
runing.movingwin.analysis(
  data = data,
  rolling.data = rolling.data,
  method = "spearman",
  myform = formula("log.seed~temperature"),
  model_type = "lm"
)
}
\arguments{
\item{data}{A data frame containing the seed count and other site-level information. It should include columns \code{Year} (or \code{year}), \code{sitenewname}, and \code{log.seed}.}

\item{rolling.data}{A data frame containing rolling climate data, including the rolling average temperature and a \code{days.reversed} column for the moving window analysis.}

\item{method}{A string specifying the correlation method to use. Default is 'spearman'. Other options include 'pearson' and 'kendall'.}

\item{myform}{A formula specifying the model to fit for each moving window. Default is \code{formula('log.seed~rolling_avg_tmean')}.}

\item{covariates.of.interest}{A string specifying the name of the climate variable to be used as a covariate in the analysis. Default is 'rolling_avg_tmean'.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{sitenewname}: The name of the site.
\item \code{plotname.lon.lat}: The plot's longitude and latitude (if available).
\item \code{days.reversed}: The days reversed for the moving window.
\item \code{term}: The term of the model.
\item \code{estimate}: The estimated coefficient for the term.
\item \code{std.error}: The standard error of the coefficient estimate.
\item \code{p.value}: The p-value for the coefficient estimate.
\item \code{r.squared}: The R-squared value of the model.
\item \code{logLik}: The log-likelihood of the model.
\item \code{correlation}: The correlation coefficient.
\item \code{pvalue.cor}: The p-value of the correlation.
\item \code{correlation.se}: The standard error of the correlation coefficient.
}
}
\description{
This function performs a moving window analysis to investigate the relationship between a response variable (e.g., seed count) and a rolling climate variable (e.g., temperature) over time. It calculates correlation coefficients, fits linear models, and extracts relevant statistics.
}
\details{
The function performs the following steps:
\itemize{
\item Merges the site-level data with rolling climate data.
\item Calculates correlation coefficients and p-values for each window.
\item Fits linear models for each moving window and extracts model coefficients and statistics.
\item Computes the standard error of the Spearman correlation coefficients.
}
}
\examples{
# Example data (here only one obs that would not work, but that's the idea)
data <- data.frame(
Year = as.numeric(2000:2019),
sitenewname = rep('site1', 20),
log.seed = rnorm(20))
rolling.data <- data.frame(
days.reversed = 1:365,
rolling_avg_tmean = rnorm(365), year = 2017)

# Run the moving window analysis
result <- running.movingwin.analysis(data = data, rolling.data = rolling.data)

}
