% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_window_ranges}
\alias{save_window_ranges}
\title{Save Window Ranges from Sequences of Days}
\usage{
save_window_ranges(sequences_days)
}
\arguments{
\item{sequences_days}{List. A list of sequences of days, where each sequence is a vector of day numbers. Each sequence represents a period of interest.}
}
\value{
A data frame with two columns:
\item{window.close}{The start day of each sequence.}
\item{window.open}{The end day of each sequence.}
}
\description{
This function processes a list of sequences of days and saves the ranges for each sequence. For each sequence, it identifies the start and end days of the window and compiles these into a data frame.
}
\details{
The function performs the following:
\itemize{
\item Checks if there is only one sequence in the list and handles it accordingly.
\item Loops through each sequence of days in the list.
\item For each sequence, it extracts the first and last day to define the range.
\item Converts the results into a data frame for easier handling.
}
}
\examples{
# Example usage of the save_window_ranges function
sequences <- list(c(1, 2, 3, 4, 5), c(10, 11, 12, 13))
window_ranges <- save_window_ranges(sequences)
print(window_ranges)

}
