% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se}
\alias{se}
\title{Calculate Standard Error}
\usage{
se(x)
}
\arguments{
\item{x}{Numeric vector. The data for which the standard error is to be calculated.}
}
\value{
A numeric value representing the standard error of the input vector.
}
\description{
This function calculates the standard error of a numeric vector, excluding missing values (\code{NA}).
}
\details{
The function first removes missing values using \code{na.omit()}. It then calculates the sample size (\code{n}), and returns the standard deviation of the non-missing values divided by the square root of the sample size (i.e., the standard error).
}
\examples{
# Example with a numeric vector
values <- c(1.5, 2.3, 3.1, NA, 4.0)
standard_error <- se(values)
print(standard_error)

}
