% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_fake_bio_data.R
\name{simulated_fake_bio_data}
\alias{simulated_fake_bio_data}
\title{Simulate Biological Data Based on Climate Windows}
\usage{
simulated_fake_bio_data(
  fakeclimatewindow,
  setup.param,
  num_simulations,
  save_fake_data = TRUE,
  overwrite = FALSE,
  save_path = here("Application_MASTREE/outputs")
)
}
\arguments{
\item{fakeclimatewindow}{A data frame of simulated climate data (must include \code{year} and \code{TMEAN}).}

\item{setup.param}{A named list with numeric ranges for parameters: \code{alpha}, \code{beta}, and \code{sigma}.}

\item{num_simulations}{Integer. Number of simulated biological datasets to generate.}

\item{save_fake_data}{Logical. Whether to save the simulations to disk. Default is \code{TRUE}.}

\item{overwrite}{Logical. Whether to overwrite existing \code{.qs} file if it exists. Default is \code{FALSE}.}

\item{save_path}{Character. Folder path where the file should be saved. Default is \code{'outputs'}.}
}
\value{
A list of data frames containing simulated biological datasets with metadata.
}
\description{
This function generates simulated biological datasets (e.g., seed production) using specified
alpha, beta, and sigma values to create controlled relationships with climate predictors.
You can save the simulated results to a specified folder using \code{.qs} format. Feel free to adjust this code for your own ;D
}
