% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATS_CSP.R
\name{ATS_CSP}
\alias{ATS_CSP}
\title{Perform Climate Signal Processing (CSP) Site Analysis}
\usage{
ATS_CSP(
  Results_days,
  siteneame.forsub,
  bio_data,
  climate.path,
  refday,
  lastdays,
  rollwin
)
}
\arguments{
\item{Results_days}{A subset of results for a specific site. Typically a data frame or list that contains climate and seed-related estimates.}

\item{siteneame.forsub}{A character string indicating the site name (longitude and latitude) used to filter the bio dataset.}

\item{bio_data}{A data frame containing biological data for multiple sites, including columns like \code{plotname.lon.lat} that are used to filter for a specific site.}

\item{climate.path}{A string specifying the file path to the climate data related for analysis.}

\item{refday}{An integer representing the reference day of the year used to align the rolling climate window.}

\item{lastdays}{An integer representing the total number of days for the climate window analysis.}

\item{rollwin}{An integer specifying the size of the rolling window for the climate data (number of days to average).}
}
\value{
A data frame or list, which is the output of the \code{runing_csp_site} function. It typically contains model estimates,
climate window ranges, and other statistical summaries related to the climate-seed analysis.
}
\description{
This function performs a CSP site analysis by filtering the provided seed data for a specific site
and then calling the \code{runing_csp_site} function to analyze climate and seed data over a specified time window.
}
\details{
This function filters the provided \code{bio_data} data by the site specified in \code{siteneame.forsub} and
then runs the \code{runing_csp_site} function with the filtered seed data, climate data, and other parameters.
The \code{runing_csp_site} function processes climate data within a rolling window to find relationships between
climate signals and seed data.
}
\examples{
\dontrun{
# Example usage of the CSP_function_site function
CSP_function_site(
  Results_days = Results_CSP[[1]],
  siteneame.forsub = "51.0_10.0",
  bio_data = bio_data,
  climate.path = "data/climate_beech",
  refday = 305,
  lastdays = 600,
  rollwin = 1
)
}

}
