% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_climwin.R
\name{climwin_site_days}
\alias{climwin_site_days}
\title{Perform Climate Window Analysis - based on climwin R package from Bailey et al}
\usage{
climwin_site_days(
  climate_data,
  data,
  site.name,
  range = c(600, 0),
  cinterval = "day",
  refday = c(1, 11),
  optionwindows = "absolute",
  climate_var = "TMEAN",
  stat.aggregate = "mean",
  formulanull = formula(log.seed ~ 1)
)
}
\arguments{
\item{data}{A data frame containing biological data including seed production and relevant metadata.
It should include columns for \code{Date2} and \code{sitenewname}.}

\item{site.name}{A character string specifying the name of the site for which the analysis is conducted.
This is used to filter the climate data specific to this site.}

\item{range}{A numeric vector of length 2 specifying the range of days over which to perform the sliding
window analysis. This defines the maximum and minimun range for window lengths in days. Defaults to \code{c(600, 0)}.}

\item{cinterval}{A character string specifying the interval at which climate data is aggregated.
Options include 'day', 'week', etc. Defaults to \code{'day'} same as climwin.}

\item{refday}{A numeric vector of length 2 specifying the reference day for absolute windows.
Defaults to \code{c(01, 11)} (e.g., November 1st) - meaning will search before 1st of November.}

\item{optionwindows}{A character string specifying the type of window to use. Options include 'absolute'
for fixed dates and 'relative' for relative to a specific event. Defaults to \code{'absolute'}. Relative might be useful
if you have a specific date time event, which is not the case with our seed production data}

\item{climate_var}{A character string specifying the name of the climate variable to analyze (e.g., temperature
mean, maximum, or minimum). Defaults to \code{'TMEAN'}. But it should basically match your same column in your climate file}

\item{climate.path}{A character string specifying the path to the directory containing
the climate data files. The function searches for files matching the \code{site.name} pattern in this directory.}
}
\value{
A data frame containing:
\itemize{
\item \code{sitenewname}: Site-specific information.
\item \code{climate.file}: Name of the climate data file used.
\item \code{climwin_output$combos}: Results from the sliding window analysis, the output from climwin::slidingwin.
\item Performance metrics of the best model (\code{performance::model_performance}).
\item Coefficients of the best model extracted using \code{broom::tidy}.
}
}
\description{
This function performs a climate window analysis using the \code{climwin} R package to identify
the best temperature windows that correlate with seed production. It processes climate and
biological data, runs a sliding window analysis, and returns statistics and windows identified
NOTE that most parameters are already well described in climwin R package ;O
You also need to specify your null model as reference, here
baseline = lm(log.seed~1, data = data)
but double check you have the proper response name variable
}
\details{
The function filters and processes climate data, performs a sliding window analysis using the \code{climwin} package,
and extracts model summaries and window. The function assumes the climate data files are in a format compatible with the \code{qs}
package (qs file are just smaller size and can be read by using qs::qread) and that the biological data contains a valid response variable as specified by \code{formulanull}.
}
\examples{
# Example usage:
# result <- climwin_site_days(
#   climate.path = "path/to/climate/data",
#   data = biological_data,
#   site.name = "site1"
# )

}
