% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_CSP.R
\name{get_predictions_windows}
\alias{get_predictions_windows}
\title{Get Window Ranges from Predictions}
\usage{
get_predictions_windows(slope_gam, rs_gam, temporary)
}
\arguments{
\item{slope_gam}{A GAM model object used for predicting slopes. This model should be fitted using \code{mgcv::gam} and provide predictions related to the response variable of interest.}

\item{rs_gam}{A GAM model object used for predicting R-squared values. This model should also be fitted using \code{mgcv::gam} and provide predictions related to the goodness-of-fit of the model.}

\item{temporary}{A placeholder parameter that is not used in the current function implementation. It is included for potential future use or consistency with other function signatures.}
}
\value{
A list containing:
\itemize{
\item \code{pred_S}: A list with predicted values and standard errors for the slope GAM model.
\item \code{pred_R}: A list with predicted values and standard errors for the R-squared GAM model.
\item \code{days}: A numeric vector representing the days where the predictions meet the specified criteria.
}
}
\description{
This function calculates prediction windows for given Generalized Additive Models (GAMs) based on slopes and R-squared values. It determines the range of days where the predicted slopes and R-squared values fall within specific quantiles, and then identifies the concurrent period with extreme values.
}
\details{
The function performs the following steps:
\itemize{
\item Predicts values and standard errors for the slope and R-squared GAM models.
\item Identifies the lower and upper range of coefficients where the predictions fall within the 2.5th and 97.5th percentiles of the fitted values.
\item Identifies the range of days where the R-squared values fall within the 97.5th percentile.
\item Finds the intersection of these ranges to determine the concurrent periods with extreme values.
}
}
\examples{
# Example usage
# Assuming slope_gam and rs_gam are previously defined GAM models
result <- get_predictions_windows(slope_gam = my_slope_gam, rs_gam = my_rs_gam, temporary = NULL)

# Accessing results
pred_S <- result$pred_S
pred_R <- result$pred_R
days <- result$days

}
