% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_csp.R
\name{runing_csp}
\alias{runing_csp}
\title{Run Climate Signal Processing for a Site}
\usage{
runing_csp(
  Results_days = Results_days,
  bio_data = bio_data,
  siteneame.forsub = siteneame.forsub,
  option.check.name = TRUE,
  climate_data = climate_data,
  refday = 305,
  lastdays = max(range),
  rollwin = 1,
  optim.k = F,
  formula_model = formula("seed~TMEAN"),
  model_type = "lm",
  yearneed = 2
)
}
\arguments{
\item{Results_days}{A subset of the \code{Results_CSP} data frame for a specific site.}

\item{bio_data}{A data frame containing site-level bio_data, typically including seed production and climate-related variables.}

\item{siteneame.forsub}{A string representing the name of the site to subset from the \code{Results_CSP} bio_data and climate data.}

\item{refday}{An integer specifying the reference day (default is 305).}

\item{lastdays}{An integer representing the last day of the range used for the analysis (default is the maximum of the range).}

\item{rollwin}{An integer specifying the size of the rolling window used for calculating temperature averages (default is 1).}
}
\value{
A data frame (\code{output_fit_summary.temp}) containing the results of the climate signal processing, including fitted model coefficients, model statistics, and window ranges.
}
\description{
This function processes climate signal data for a specific site by fitting a generalized additive model (GAM) and running window-based modeling.
}
\details{
The function processes climate signal data for a specific site by running the following steps:
\itemize{
\item It extracts slopes and R-squared values from \code{Results_days} and constructs a temporary data frame for further analysis.
\item It fits a generalized additive model (GAM) to the slopes and R-squared values using the \code{optimize_and_fit_gam} function.
\item It identifies consecutive sequences of significant days (\code{extract_consecutive_sequences}) and generates window ranges.
\item It loads the climate data for the site from the specified path, formats the dates, and scales the climate variables.
\item It computes rolling temperature data for the selected time period.
\item It applies the \code{reruning_windows_modelling} function across all identified window ranges to model the relationship between seed production and climate variables.
}
}
\note{
The function assumes that the \code{bio_data} argument corresponds to the site-level data and contains a column named \code{plotname.lon.lat} for site identification.
}
\examples{
\dontrun{
# Example usage:
output <- runing_csp_site(Results_days = Results_CSP[i],
                          bio_data = bio_data,
                          siteneame.forsub = "site_123",
                          climate.path = "/path/to/climate/data")
}

}
