% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_transformation_betareg.R
\name{y_transformation_betareg}
\alias{y_transformation_betareg}
\title{Transform Response Variable for Beta Regression}
\usage{
y_transformation_betareg(y)
}
\arguments{
\item{y}{A numeric vector of response values that are to be transformed. The vector
can contain NA values which will be ignored in the calculation of the transformation.}
}
\value{
A numeric vector of the same length as \code{y}, containing the transformed response values.
Values are scaled to the range (0, 1), suitable for beta regression.
}
\description{
This function transforms a response variable \( y \) for use in beta regression
by scaling it to fit the range required by beta regression models. The transformation
is performed as follows:
}
\details{
\deqn{y' = \frac{y \cdot (n - 1) + 0.5}{n}}

where \( n \) is the number of non-missing observations in \( y \).
}
\examples{
# Example usage
y <- c(0.1, 0.5, NA, 0.3, 0.9)
transformed_y <- y.transf.betareg(y)
print(transformed_y)

}
